fix(misra): cast unsigned clamp minimum to target type and suppress floatConversionOverflow in MISRA runner (issue #990)
Two root causes for MISRA failures in float-cast unsigned clamp tests:

1. CodeGenerator.generateFloatToIntClampCast() used bare '0' for unsigned
   type minimums instead of casting to the target type (e.g. (uint8_t)0).
   This violated MISRA C:2012 Rule 10.4 (operands must have same essential
   type category). Now uses CppModeHelper.cast() uniformly for all types.

2. batch-validate.mjs MISRA runner lacked --suppress=floatConversionOverflow
   that the cppcheck runner already had. cppcheck cannot reason through
   ternary guards in clamp expressions, so it false-positives on guarded
   casts like (uint8_t)(over) even though the ternary prevents overflow.

+++ tests/floats/float-int-conversion.test.cpp
@@ -27,7 +27,7 @@ void test_explicit_cast(void) {
     float a = 3.14;
     int32_t truncated = ((a) > ((float)INT32_MAX) ? static_cast<int32_t>(INT32_MAX) : (a) < ((float)INT32_MIN) ? static_cast<int32_t>(INT32_MIN) : static_cast<int32_t>((a)));
     double b = 99.99;
-    uint32_t unsigned_val = ((b) > ((double)UINT32_MAX) ? static_cast<uint32_t>(UINT32_MAX) : (b) < 0.0 ? 0 : static_cast<uint32_t>((b)));
+    uint32_t unsigned_val = ((b) > ((double)UINT32_MAX) ? static_cast<uint32_t>(UINT32_MAX) : (b) < 0.0 ? static_cast<uint32_t>(0) : static_cast<uint32_t>((b)));
     double big = 3.141592653589793;
     float small = static_cast<float>(big);
 }
